India Covid Deaths
This is my first try at a Machine learning project using the Linear Regression Algorithm and an India Covid dataset. This model predicts the deaths due to covid based on the total cases, active cases, and discharged cases. Currently the project outputs the test results of a 20% data set test which shows what the input data is as well as what the model predicts the covid deaths will be. The correct result is also printed on screen to allow a user to compare predictions with correct results to check for accuracy. A scatter plot is also displayed to the sure which shows the correlation between covid deaths and total cases. The project is written in python using Pycharm IDE. The project uses sklearn, numpy, pandas, pickle, and matplotlib.