Water_Temp_ML
For this project, I decided to use a dataset which contained various data on water to answer the question of whether salinity of water had any correlation to water temperature. I coded this ML model using python and the Pycharm IDE. I wanted to practice the linear regression algorithm and thought this was a good use for it. I used pandas, numpy, and sklearn for the linear_model. I spent many hours cleaning out the data because it contained many rows which had empty data which was either causing type errors or i felt like would incorrectly skew data. After doing so, i passed this data inputs/ outputs to train my model. initially I had 78 pieces of data which I trained my model on and found that the model would return anywhere from 70% to 75% accuracy, however i wanted to train my model with more data I had available. On the next run, I added a total of 8k data to train the model and all of a sudden the accuracy dropped to 40%. I added a for loop to train the model 5000 times and get the average accuracy from all the scores which was 39%. I increased the data to 10k and then 12k and the average success rate stayed at 39%. After writing some code to display what the data being fed to the model was and its prediction compared to the correct output, I noticed that it was having a really hard time predicting the waters temperature. Sometimes it would guess double digits when the correct output was low single digits. My guess is that the original run with 78 pieces of data had a higher success rate because the model did not have enough data to make a reliable prediction. I also have come to the conclusion that salinity in water cannot help the model predict a waters temperature. I decided to expand my research to include the depth of water in meters, I did the exact same process mentioned before but this time I trained it with over 12k pieces of data. The prediction accuracy average became 67.5% with both salinity and water depth. leading me to believe that water depth was a much better piece of data to predict water temperature. I finally tested to see what would happen if I removed the salinity data all together from the equation, the average percentage stayed at 67.5%. This was a good project for me because it introduced me to doing basic data cleaning. I also played around with what data affected the predictions and noticed that some date can help the model while others don't or can hurt the reliability. I assume that if I added more data from the dataset for the model to train like, Milliliters of dissolved oxygen per Liter seawater, or Potential Density of Water, among others the models' accuracy and reliability would increase provided its trained with the proper data.