WW2_Weather_MinMax_Temp
This small project started as a question on whether the minimum temp on a day could help predict the max temp of that same day. I used a WWII dataset that had various pieces of data including max/min temps for a day. I decided to use Linear Regression once again as I wanted more practice with this algorithm and it seemed like a good fit. This is because if there is indeed some sort of correlation between a days max/min temp, the algorithm should be able to find a reliable best fit line on the data. After getting the necessary data and cleaning any incompatible types from it, I created the linear regression model using python and the Pycharm IDE. In terms of libraries, i used pandas, sklearn for the algorithm, numpy for the input arrays, and matplotlib to create a visualization of the data in a graph. I trained the model with 80% of the dataset and the other 20% was reserved for testing. After training, I tested the model multiple times with an average of 76% accuracy rate. I also wrote a for loop to display all of the models predictions with the correct output as well. Visually analyzing this data shows that the model most times was off in the correct max temp degrees anywhere from 1c to 3c. This to me seems relatively close and shows that the model can get close to guessing a temperature often. This model could probably be improved by removing anomalies from the data set. For fun i also tested adding one more piece of data from the dataset to train the model with, that is, the mean temperatures. With the mean temperatures, the models accuracy improved by about 20%. Visually analyzing does results also showed that the predictions done by the model where often times almost identical to the real correct results. As stated before, I used matplotlib to create a graph which showed the data plotted of maxtemp(y) and mintemp(x) this graph clearly shows that as the minimum temp goes up so does the maximum temp go up for any particular day. This was another great lesson in machine learning and linear regression for me.