Relationship between humidity and temp
This small linear regression project was made in order to create a model which could show the relationship between humidity and apparent temperature and humidity and temperature. The libraries used for this project were, pandas to load the dataset being used. sklearn using linear_model to create the actual model and train/test it, numpy to create useable train and test data, and matplotlib to show 2 scatter plot graphs to demonstrate any relationship visually. After completing this project the scatter plots show that there is no concrete relationship between humidity and temp/apparent temp as seen by how random the data is between x and y axis.